[CLI] devide main logic into modules
